Delete the file once we have finished processing it
This means we won't needlessly waste disk space. This actually
isn't a problem on Heroku, as it runs on multiple dynos (and wipe
the hard drive of each dyno when it goes to sleep), so the files
will eventually be deleted automatically. However if we want to
migrate the codebase to another web hosting service, like AWS,
this code will probably be essential.